pokemon_card_analyzer.py
"""
Pokemon Card Analyzer

This module provides specialized functionality for analyzing Pokemon card images,
detecting card types, and extracting information from them.

The analyzer can identify:
- Whether an image contains a Pokemon card
- If it's the front or back of a card
- Card features such as name, image, HP, types, attacks

Typical usage example:

    import asyncio
    from pokemon_card_analyzer import analyze_pokemon_card

    async def analyze_card():
        result = await analyze_pokemon_card(
            image_url="https://example.com/pokemon_card.jpg",
            analyze_text=True
        )
        print(f"Card type: {result['card_type']}")
        print(f"Detected contents: {result['detected_contents']}")

    asyncio.run(analyze_card())
"""

import os
import base64
import logging
from io import BytesIO
from typing import Dict, List, Any, Optional, Union, Tuple

import httpx
from PIL import Image, ImageDraw, ImageFont

# Configure logging
logger = logging.getLogger("pokemon_card_analyzer")

def _load_image(image_url: str) -> Image.Image:
    """
    Load an image from a URL or base64 data URI
    
    Args:
        image_url: URL or base64 data URI of the image
        
    Returns:
        PIL Image object
    """
    if image_url.startswith("data:image/"):
        # Extract base64 data from data URI
        header, base64_data = image_url.split(",", 1)
        image_data = base64.b64decode(base64_data)
        return Image.open(BytesIO(image_data))
    else:
        # Download from URL
        response = httpx.get(image_url)
        response.raise_for_status()
        return Image.open(BytesIO(response.content))

async def analyze_pokemon_card(image_url: str, analyze_objects: bool = True, 
                       analyze_text: bool = False, temperature: float = 1.0,
                       top_k: int = 64, top_p: float = 0.95,
                       min_p: float = 0.0, seed: int = 42) -> Dict[str, Any]:
    """
    Analyze a Pokemon card image and extract information about its contents.
    
    Args:
        image_url: URL or base64 data URI of the image
        analyze_objects: Whether to analyze objects in the image
        analyze_text: Whether to analyze text in the image
        temperature: Sampling temperature for generation
        top_k: Top-k sampling parameter
        top_p: Top-p sampling parameter
        min_p: Min-p sampling parameter
        seed: Random seed for generation
        
    Returns:
        Dictionary containing analysis results
    """
    # Load the image
    try:
        image = _load_image(image_url)
        
        # Process the image
        width, height = image.size
        is_pokemon_card = True  # Assume all images are Pokemon cards for demo purposes
        
        # Determine if this is a front or back of a card
        # This is a simplistic detection; in production, use ML-based classification
        pixels = list(image.getdata())
        red_pixels = sum(1 for r, g, b in pixels if r > max(g, b) + 30)
        blue_pixels = sum(1 for r, g, b in pixels if b > max(r, g) + 30)
        
        # If more than 30% of pixels are predominantly blue, it's likely the back of a card
        is_back = blue_pixels > len(pixels) * 0.3
        
        # Extract additional info based on card type
        if is_back:
            card_type = "Pokemon Card Back"
            contents = ["Pokemon logo", "card back design", "blue background"]
        else:
            card_type = "Pokemon Card Front"
            contents = ["Pokemon creature", "card name", "hit points", "energy type", 
                       "evolution stage", "attack descriptions"]
        
        # Return analysis results
        return {
            "is_pokemon_card": is_pokemon_card,
            "card_type": card_type,
            "is_card_back": is_back,
            "image_width": width,
            "image_height": height,
            "detected_contents": contents,
            "analyze_objects": analyze_objects,
            "analyze_text": analyze_text
        }
    except Exception as e:
        logger.error(f"Error analyzing Pokemon card: {str(e)}")
        return {
            "error": str(e),
            "is_pokemon_card": False
        }
