pyproject.toml
[project]
name = "gemma3"
version = "0.1.0"
description = "OpenAI-compatible server for Gemma 3 with function calling support"
authors = [{name = "AI Developer"}]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "llama-cpp-python>=0.2.11",
    "pydantic==2.11.4",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.2",
    "psutil>=5.9.0",
    "typing-extensions>=4.8.0",
    "httpx>=0.28.1",
    "pytest-asyncio>=0.26.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pygame>=2.6.1",
    "requests>=2.32.3",
    "python-multipart>=0.0.20",
    "urllib3>=2.4.0",
    "honcho>=2.0.0",
    "attrs==25.3.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2025.4.1",
    "referencing==0.36.2",
    "rpds-py==0.25.0",
    "pydantic-core==2.33.2",
    "pillow>=11.2.1",
]
requires-python = ">=3.9"

[project.optional-dependencies]
test = [
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pytest-asyncio>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=gemma3"
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"

[project.scripts]
gemma3 = "gemma3:main"
